create a default-constructed object when calling native constructors
It's up to the implementation of the constructor to decide whether
it wants to initialize the default-constructed object, or create
and return a new one; just like for JS constructors.
We need this functionality in QtScript so that
QScriptContext::thisObject() will return the right thing.
The built-in constructors (RegExp etc.) should be modified to
initialize the default-constructed object instead of always
creating a new object (only when RegExp is called as a function
should it create a new object).