get rid of some expected failures
The JSC parser doesn't understand function expressions as
statements; you have to surround the expression by parentheses.
That's in accordance with the ECMA spec, but most
implementations (including the old qtscript backend) don't
require the parentheses. However, since it's easier to change
the tests than the JSC parser, let's do that for now to get
rid of some noise in the test output.