QGraphicsAnchorLayout: Do not create center anchors by default
Previously we would create two half anchors internally to each
item added to the layout. These anchors are meant to enforce the
central anchorage point remains on its position.

It is not effecient however, to create these anchors for all
items if not all of them will have anchors linked to their center.

This commit removes the creation of center anchors, the idea is
to postpone this to the moment their are really needed, therefore
reducing the number of variables and constraints sent to the simplex
solver on the average case.

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>
Reviewed-by: Anselmo Lacerda S. de Melo <anselmo.melo@openbossa.org>