Removed an assert from the QNetworkReply.
It is possible that finished() function will be called several times,
for example if the user calls QNetworkReply::close() as a result of
the authenticationRequired() signal, the connection will be closed
(which results in emitting a finished() signal), and then the
authentication will be cancelled, which will also try to emit the same
signal, however the connection state will already be set to Finished.

The fix is to just make sure if the finished() signal has already been
emitted by checking if the connection state is not Finished or
Aborted.

Reviewed-by: Thiago Macieira