add some world-class hacks to JSC to make QtScript work
QtScript needs the VM to create script objects of type
QScriptObject, not JSC::Object. This is so that the class of
the object can be changed dynamically using e.g.
QScriptValue::setScriptClass(), or the overload of
QScriptEngine::newQObject() that takes an existing object
as first argument (it "promotes" the plain script object to
a QObject wrapper).

This makes the bindings produced by the bindings generator
work.

When JSC becomes a shared library, we need another solution.