Fixed a bug where the input panel would not show up.
Task:     257214

The problem happened if a user called setFocus() on an input capable
widget, and then tried to open the input panel by sending an event.
Since the call to InputCapabilitiesChanged is asynchronous, Symbian
would not yet know about the updated state, and the event would be
lost.

Now we generate our own asynchronous event, and ensure that it is
synchronous in the cases where it's needed.