Cut-off in QGraphicsView not hit if setting a scene rect.
QGraphicsScene::sceneRect() returns the growingItemsBoundingRect if not
a particular scene rect is set; otherwise the sceneRect is returned.
This cut-off aims to do less processing when the exposed region covers
the entire scene, but it won't get hit in case of a sceneRect because we
always check against the growingItemsBoundingRect.

Task-number: 257192