Fix font propagation issues with QComboBox and the popup menu.
This has always been a bit bumpy, the problem is that the popup normally
has its own styling from the desktop, plus it's its own top-level and
that is normally a boundary for propagation. Of course, people are
surprised by this (especially when it works for editable). So, we need
to be a bit better propagating the info. Also the QStyleOptionMenuItem
has the correct font, but if it's set on a window, by the time it
reaches the popup, its resolve mask is very weak, so it will fail to
resolve at all. Setting the point size allows the font to have a bit of
strength.

Task-number: 257486
Reviewed-by: Jens Bache-Wiig