Bad drawing of styled viewports within QAbstractScrollArea
This patch includes lots of refactoring, but the real problem was that
in QWidgetPrivate::paintBackground we call drawPrimitive(PE_Widget) with
a potentialy translated painter,   but the opt.rect is not translated.

When having a scroll area the calling function used to translated the
painter and then pass the offset around to rectify. but drawPrimitive
cannot rectify it.

The solution is not to translate the painter but use other way to
rectify the brush

Task-number: 257517
Reviewed-by: bnilsen