Use Unix signal handlers to properly close the test log when a test is terminated by a signal.
The new XML testloggers in Qt 4.6 only output a testlog after the test
completes.  Therefore, if a test crashes, no testlog will be output at
all.  This is a functional regression from Qt 4.5 and earlier where the
testlog would be output up to the point where the crash occurred.

This is a Unix-specific fix for this problem.

This change is also useful for hanging tests; if the test is killed
with SIGTERM it will output the test log up to the current test before
it exits.