Fixed a consistency problem in the backing store.
Widgets maintain an internal state (inDirtyList) which says whether
that widget is in the backing store's list of dirty widgets.
Naturally this list has to be kept in sync with the actual list
contents. This did not happen if the backing store was destroyed and
recreated; if the list had any contents it would be destroyed without
resetting the widget flags. Destroying and recreating the backing
store is common on S60 where we try to save memory.

Fixed by making sure those flags are reset during backing store
destruction.

Done together with Jason.
RevBy:    Jason Barron
RevBy:    bnilsen
AutoTest: Included