Make the mainwindow examples use setWindowFilePath
Should have done this a while ago. Use setWindowFilePath
to get proxy icons on mac and simplify window
title handling on the other platforms. The only
downside is that we do have to specify the
application name to get the exact same behavior.
But on the other hand, QSettings handlig is a bit
better.