Make the window title set by setWindowFilePath() match our docs.
On the mac, we were setting the absolute path of the file passed in
while we document that we only set the filename. The filename is
actually what all the other Mac apps do, so we should really do it.

Unfortunately, this task laid around longer than it should, but the
description text was bad.

Additionally, fix a bug in the Cocoa version where we would set a proxy
icon for items that didn't exist, this is not a good idea.

Task-number: 220501
Reviewed-by: Trust Me