Fixed text rendering with the GL2 paint engine.
When copying a glyph image into the glyph cache, garbage appeared in
the glyph cache in the lower part of the destination rectangle. This
happened whenever the glyph image's width was not a multiple of four
bytes. I suppose this is a driver bug (nVidia). As a workaround, I
converted the glyph image to ARGB32 such that the width is guaranteed
to be a multiple of four bytes.

Reviewed-by: Tom