Fix configure.exe when more than 256(!) else-ifs are required.
The introduction of the additional 'continuousElse' statement was
flawed because once the variable is toggled to true by the first else
block, it remains true for the second block even if the option was
found so these options will eventually hit the "Unknown option" error
in the final else case.

The workaround is not to re-use the variable for each block, but
instead use an array of bools. One element of the array per
'continuousElse' check. This should probably be refactored soon-ish :)