Better caching of file system icon providers.
It turns out that we weren't doing any caching of icons provided by the
file system. We now use the similar trick that's used on Windows which
does some caching on the file extension. We do fill up the cache
needlessly with extra information (16, 32, 64, and 128) icons. We
probably could be better with a iconRef engine that generates these
sizes on demand. Still performance is 100% better with this which means
using it in itemviews works.

Reviewed-by: Jens Bache-Wiig