QDirIterator: reducing "randomness"
The difference between a canonical and absolute paths is subtle, and not
what QDirIterator is about. With this change, we still avoid loops
generated by symbolic links but won't duplicate entries because of these
differences.

While at it, when avoiding loops with symbolic links, please don't mess
with the next path! That only added inconsistency.

Reviewed-by: Olivier Goffart