Increased default stack size for QThreads on Symbian.
It can be debated whether this is a bug or not, and whether or not we
should rely on the Symbian default. Our rationale was that the
default size caused even the simplest tasks to fail (the task talked
about a file copy crashing), so we decided to increase the headroom.
Those that want to minimize the stack size can do so manually with
setStackSize().

We go by the old API design slogan: Make the simple tasks easy, and
the advanced tasks possible.

RevBy:    Jason Barron
Task:     253666