Get collapsible menus working correctly.
There was an attempt to do this earlier, but it was a bit more complex
than it needed to be. We now do the update on show in Cocoa. Carbon
actually does it all for us, we just need to flip the bit. We may do the
updates to often, but it's better than not enough.

Task-Id: 195445
Reviewed-by: Denis