Refactoring QDirIteratorPrivate::pushSubDirectory
pushSubDirectory was operating on nextFileInfo when it should really be
using the path received as argument.

This fixes an issue introduced when currentFilePath variable was
removed, that was exposed in the auto-tests; fixes a regression
introduced in 4.5.0 -- test case a couple of commits back.

This also allows refactoring calling code and avoid repetition.

Task-number: 258230
Reviewed-by: Olivier Goffart