Allow picking up tool button style from the system
KDE and GNOME has the concept of a default tool button style that
can be set system-wide. Qt currently allways default to IconOnly.
I have added an optional Qt::ToolButtonSystemDefault value so you
can opt-in to respect the system setting. We did not change the
default because a lot of apps will look odd when for instance
text is beside icons and the descriptive text is too long.

Task-number: 237864
Reviewed-by: ogoffart