QDirIterator: no point in delaying the inevitable
The only reason I see for not calling advance() directly in the
constructor is to delay potentially unnecessary work.

However, since copy constructors have been explicitly disabled, once
QDirIterator is instantiated one has signed up for all the pain that
comes with it.

That's also a couple less conditionals in each iteration of normal use
cases.

Reviewed-by: Marius Storm-Olsen