Get monotonic time working on Mac OS X for corelib programs.
Mac OS X does not provide POSIX monotonic timers. Instead it does
provide a Mach call to get the absolute time (a.k.a., number of CPU
ticks) for the next timer event. This gets us around the bug in
select(2) on Mac, that it doesn't wakeup when the times been changed.

Of course, if you used the GUI event dispatcher, which is based on
CFRunLoopTimers, this is not an issue, but if you really just need
corelib, it's a bear to bring in the other stuff.

Thanks to the nice guys at Parallels for the basics of the patch!

Task-number: 237384
Reviewed-by: Bradley T. Hughes