LayeredPane should not be reported as an IP address edit control to MSAA
The reason was that ROLE_SYSTEM_IPADDRESS = 0x3F has been added to MSAA
at one point in time. (Can be found in recent versions of OleAcc.idl).
Since the MSAA bridge used a direct mapping between QAccessible::Role
and MSAA roles this lead to that LayeredPane was interpreted to be an
IP address edit control, affecting QStackedWidget (and some relatives).
This caused some screen readers to be confused when the
same accessible interface had children such as push buttons. I also
discussed this change with Harald.

Task-number: 257958