Revert "Added a check that X11 timestamp goes forward only."
In some cases we might get an invalid timestamp that is far away in
the future, so remembering it will break all consequent X calls that
require a timestamp because it just contains junk (for example
clipboard will stop working). This happens with XIM+SCIM pair -
whenever we start input method and type something to the widget, we
get a XKeyPress event with a commited string, however the 'serial' and
'time' members of the XEvent structure are not initialized (according
to valgrind) and contain junk.

This reverts commit 2ed015b8a0ffad63f0f59b0e2255057f416895fb.

Reviewed-By: Brad