Fix handling of invalid object paths and signatures in release mode.
I had this #ifdef __OPTIMIZE__ there so that the compiler would know
not to generate unnecessary calls and a long jump table for the switch
of the marshalling code.

Turns out that in release mode, the checks I added to make sure we
detect invalid object paths and signatures were never hit (we always
treated them as pure strings).

So use the signature- and object path-checking code in both release
and debug mode.

Task-number: reported via email (tst_qdbusmarshall failing)
Reviewed-by: Peter Hartmann