Fix dfbwindowsurface handling for offscreen mode
This patch vastly simplifies the geometry handling (setGeometry/move)
It also implements a mode in which DirectFB implementations that do not
support windows can use an offscreen buffer as its backing store.
Previously the only way to do this was to paint directly on the primary
surface. This didn't work when the dfb driver didn't support an
accelerated mouse cursor.

It also detects the situation when the cursor isn't accelerated and
takes care of painting it manually when needed.

Reviewed-by: Donald <qt-info@nokia.com>