Fix two errors of QDirIteratorPrivate::matchesFilters()
The filter for includeSystem was exectuted twice. The second time
was not correct according to d9a620633d0a5fa5e69ab06ec9a706118f3df2a6
(QFileInfo::exists() can return false for system file).

For skipDirs, a parenthesis was missing in the test of includeHidden
and includeSystem. This was introduced in the refactoring of
44766d265c16551043d2739171069fe042c40091

Reviewed-by: João Abecasis