Fix QSystemTrayIcon causing three activated signals on doubleclick
The problem was that on Windows, we would activate on WM_LButtonUp,
but a double click after activating will also generate a second
WM_LButtonUp. Hence we get three activations. The fix was basically
to filter out the second WM_LButtonUP, something we also do in
qapplication_win.cpp.

Task-number: 205499
Reviewed-by: denis