Sun CC 5.9 still doesn't support Template-Template Parameters
I had added the version check when we only had CC 5.5 and 5.6,
expecting that 5.7 would have the support. And if it didn't, then
someone would notice the compile error in QtConcurrent, bumping the
version number here.

Except that QtConcurrent was never enabled with Sun CC. Which meant
that we never got to test TTP support.

Reviewed-By: Bradley T. Hughes