Work around a Sun CC 5.9 compiler bug: the threadEngine variable isn't found.
QtConcurrent had the following code:
template <typename T> class ThreadEngineStarterBase
{ ...
protected:
    ThreadEngine<T> *threadEngine;
};

template <typename T>
class ThreadEngineStarter : public ThreadEngineStarterBase<T>
{
public:
    ThreadEngineStarter(ThreadEngine<T> *threadEngine)
    :ThreadEngineStarterBase<T>(threadEngine) {}
[...]
};

The Sun CC compiler simply didn't parse the parameter declaration in
the constructor. Instead of complaining, it silently ignored the
problem. Which meant that the constructor simply used the
uninitialised member variable from the base class to call the parent
constructor, which ended up initialised with itself.

You'd think that it's just because the parameter has the same name as
a member variable. But it appears to be a compiler bug altogether. If
you change the name, then you start getting compile errors.

This change is a workaround that worked.

Reviewed-By: Bradley T. Hughes