Re-disable QtConcurrent build with Sun CC 5.9.
QtCore compiled, but of course that means very little since most of
QtConcurrent's problems are in the template code, which isn't
instantiated inside QtCore. Examples and tests all failed to build...

I'm almost done getting it to work with that compiler, but it will
take me a little more time. I'm refactoring a bit of the QtConcurrent
code, so that things compile more smoothly, using partial template
specialisation.

To be on the safe side, re-disable for xlC 7 too.