Remove the old specialised deleter implementation from QSharedPointer.
This should be binary- and source-compatible, since these
QSharedPointer internal classes aren't exported. The compiler should
generate the symbols in all libraries and applications that used it,
which means removing it from Qt won't affect them.

(In fact, these symbols shouldn't be in QtCore at all, since we don't
use QSharedPointer in it)