Adapt the contiguous-creator code to use the new custom deleter code.
We use the same trick as the custom deleter: we allocate memory for an
object of class ExternalRefCountWithContiguousData<T>, but we do that
only to be certain of the alignment requirements for T. We initialise
the d-pointer via placement new and the T object is left for
initialisation by the outermost function.

The reason for that last trick is to support passing parameters in the
future with the least amount of template functions necessary. I still
plan on supporting arguments only with C++0x (maybe up to one
without).