Fixed failure of tst_qlocale if run when LC_ALL is set.
This test tried to be unaffected by the locale environment variables
set by the user by explicitly overwriting them and spawning a new
process.  However this was only done for LANG and not the other
variables which can affect the locale.  Do it for all the locale
variables.

Also, the previous approach (store the current value with qgetenv, then
put it back with qputenv later) doesn't really work because that
effectively takes variables which weren't set and sets them to an empty
string, which is not the same thing.  Use QProcess::setEnvironment
instead.

Reviewed-by: Andy Shaw