Roll back the fancy updating of translucent windows.
The problem with the fix, though it produces less flicker when
resizing, is that it delays telling windows that the window has moved
until after the window has been completely repainted. Problem with
this is that functions that rely on windows to be up to date will fail
until the backbuffer is flushed. This was the case for mapTo/FromGlobal, and potentially other functions too.

Reviewed-By: Eskil