Don't animate widgets on first show in Vista
Checkboxes would animate when first shown if they were checked.
This is unintentional and looks a bit odd in wizard for instance.
To fix this we simply check if the old state was set. Note that
this is safe because we will at least require the enabled state
flag to be set.

Task-number:253075
Reviewed-by: ogoffart