Implement a new pointer-tracking mechanism for QSharedPointer.
Some compilers don't obey the same rules of "top-of-object" values for
casting a pointer from a given class to void *. In any case, that can
only work for polymorphic types (with a virtual table).

So don't track the pointers by their pointer value, but instead by the
d-pointer of the QSharedPointer object.

The same cases that were caught before should still be caught. We
still won't catch the creating a second QSharedPointer for the same
object if the pointer values are different, though (when cast to
void*).

Reviewed-by: Bradley T. Hughes