Fix remaining autotest failures in tst_QGraphicsWidget
Change f68fed3 introduced a few regressions in the QGraphicsWidget
autotests. It turned out those autotests relied on behavior that this
fix "fixed". The exact bugs were 1) that setting focus on a window
or a child of a window that isn't active will automatically give that
item focus, despite that its window is inactive (in contrast it should
just set up subfocus and give the item focus when the window is
activated), and 2) that adding a window to a scene that is active did
not immediately activate that window.

So one fix in the test and one in QGraphicsScene. The autotests were
modified so that the respective tests operate on an active scene (by
assigning the scene to an active view). The change in QGraphicsScene
ensures that the first window that gets added to an active scene that
does not have any active windows already, automatically gets activated.

Reviewed-by: Michael Brasser