Ensure hover enter events are dispatched on mouse press.
This change ensures that mouse presses received by the scene when there
are no current mouse grabbers trigger hover event delivery. This is
useful when the scene only receives presses, and no mouse moves (e.g.,
disabling mouse tracking on the viewport, or on systems where the
mouse press is the first received event).

Reviewed-by: Michael Brasser