Move QLineEdit logic into QLineControl
Manually merged from the kinetic repository, where it has been used
for a while. Like QTextControl, there is no auto test for the control,
just for the widgets built with it.

The purpose of this change is so that we can build stuff like QLineEdit
in other places (like kinetic). Also included is a trivial change to
QValidator, needed to use it with QLineControl in QML.

Reviewed-by: mbm