Enter a scope when enterning a native function.
so native function that would call
engine->evaluate("var b = 'foo');
would not change the global object.

The change in qscriptengine.cpp makes sure that the correct scope is
used for the execution of QScriptEngine::evaluate.

The changes in qscriptfunction.cpp push a new scope for native function
calls.  We might want to move that into QScriptContext later

Reviewed-by: Kent Hansen